move to per function files with default exports
consistent with the rest of the library's style, in line with suggestion by @TrySound
