+++ src/index.js
@@ -3,7 +3,11 @@ export { default as v3 } from './v3.js';
 export { default as v4 } from './v4.js';
 export { default as v5 } from './v5.js';
 export { default as NIL } from './nil.js';
-export { default as validate, isV1, isV3, isV4, isV5 } from './validate.js';
 export { default as version } from './version.js';
+export { default as validate } from './validate.js';
+export { default as isV1 } from './isV1.js';
+export { default as isV3 } from './isV3.js';
+export { default as isV4 } from './isV4.js';
+export { default as isV5 } from './isV5.js';
 export { default as stringify } from './stringify.js';
 export { default as parse } from './parse.js';
