+++ src/isV5.js
@@ -1,0 +1,7 @@
+import validateByVersion from './validate-by-version.js'
+
+function isV5(uuid) {
+    return validateByVersion(uuid, 5);
+}
+
+export default isV5;
